// src/app/shared/models/macrocilo/macrociclo.model.ts
export class MacroCiclo{
    ciclo: number;
    clienteAtendido: number;
    dias: number;
    fin: string;
    flete: number;
    idArea: number;
    ingrXKm: number;
    inicio: string;
    kms: number;
    kmsCargados: number;
    kmsCargadosPor: number;
    kmsVacios: number;
    kmsVaciosPor: number;
    operacion: string;
    operadorCiclo: number;
    ruta: string;
    udN: string;
    viajes: number;
    viajesCargados: number;
    viajesVacio: number;
}